pyproject.toml
[tool.poetry]
name = "multitenancy-fastapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "multitenancy_fastapi"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
sqlalchemy = "^2.0.19"
python-multipart = "^0.0.6"
alembic = "^1.11.1"
psycopg2 = "^2.9.6"
bcrypt = "^4.0.1"
passlib = "^1.7.4"
python-dotenv = "^1.0.0"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
